requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
google-adk

